+++ calculator/src/main/java/xyz/dbotfactory/calculator/model/DebtReturnTransaction.java
@@ -2,6 +2,8 @@ package xyz.dbotfactory.calculator.model;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Builder
 @Data
 @NoArgsConstructor
@@ -9,5 +11,5 @@ import lombok.*;
 public class DebtReturnTransaction {
     long fromId;
     long toId;
-    double amount;
+    BigDecimal amount;
 }
